+++ src/mcmc/Generator.cpp
@@ -45,7 +45,7 @@ float Generator::gaussByDoubleExp(const float mean, const float sigma, const flo
     }
     
     float x = (x_max + x_min) / 2;
-    const float sqrt_e = sqrt(exp(1.));
+    const float sqrt_e = sqrtf(expf(1.));
     
     float ur = 1;
     float rap = 0;
@@ -56,27 +56,27 @@ float Generator::gaussByDoubleExp(const float mean, const float sigma, const flo
         
         if(x_min < 0 && x_max > 0)
         {
-            const float c = 1 - 0.5 * (exp(x_min) + exp(-x_max));
-            const float f0 = 0.5 * (1 - exp(x_min)) / c;
+            const float c = 1 - 0.5 * (expf(x_min) + expf(-x_max));
+            const float f0 = 0.5 * (1 - expf(x_min)) / c;
             
             if(u <= f0)
             {
-                x = log(exp(x_min) + 2*c*u);
+                x = logf(expf(x_min) + 2*c*u);
             }
             else
             {
-                x = -log(1 - 2*c*(u-f0));
+                x = -logf(1 - 2*c*(u-f0));
             }
         }
         else
         {
             if(x_min >= 0)
             {
-                x = -log(exp(-x_min) - u * (exp(-x_min) - exp(-x_max)));
+                x = -logf(expf(-x_min) - u * (expf(-x_min) - expf(-x_max)));
             }
             else
             {
-                x = log(exp(x_min) - u * (exp(x_min) - exp(x_max)));
+                x = logf(expf(x_min) - u * (expf(x_min) - expf(x_max)));
             }
         }
         
@@ -91,7 +91,7 @@ float Generator::gaussByDoubleExp(const float mean, const float sigma, const flo
         {
             if(x_min > 1)
             {
-                rap = exp(0.5 * (x_min * x_min - x * x) + x - x_min);
+                rap = expf(0.5 * (x_min * x_min - x * x) + x - x_min);
                 
                 /*std::cout << "------" << std::endl;
                 std::cout << "floatExp : x_min : " << x_min << std::endl;
@@ -105,11 +105,11 @@ float Generator::gaussByDoubleExp(const float mean, const float sigma, const flo
             else if(x_max < -1)
             {
                 //std::cout << "floatExp : x_max : " << x_max << std::endl;
-                rap = exp(0.5 * (x_max * x_max - x * x) + x_max - x);
+                rap = expf(0.5 * (x_max * x_max - x * x) + x_max - x);
             }
             else
             {
-                rap = exp(-0.5 * x * x + abs(x)) / sqrt_e;
+                rap = expf(-0.5 * x * x + abs(x)) / sqrt_e;
             }
         }
         //ur = randomUniform();
@@ -123,7 +123,7 @@ float Generator::boxMuller()
 {
     float rand1 = randomUniform();
     float rand2 = randomUniform();
-    return sqrt(-2. * log(rand1)) * cos(2. * M_PI * rand2);
+    return sqrtf(-2. * logf(rand1)) * cos(2. * M_PI * rand2);
 }
 
 float Generator::gaussByBoxMuller(const float mean, const float sigma)
